oracle-2/services/ai-router/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
slowapi>=0.1.8
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
orjson==3.10.7
PyJWT[crypto]>=2.8.0
numpy>=1.24.0
Pillow>=10.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

